backend/pyproject.toml
[tool.poetry]
name = "expense-manager"
version = "0.1.0"
description = "Expense Management System API"
authors = ["generated"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.23.0"
sqlmodel = "^0.0.8"
passlib = "^1.7.4"
python-jose = "^3.3.0"
httpx = "^0.24.0"
alembic = "^1.11.1"
python-multipart = "^0.0.6"
Pillow = "^10.0.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
backend = "poetry.core.masonry.api"